sort pom file using transform
