+++ generators/maven/support/index.ts
@@ -19,4 +19,5 @@
 
 export * from './dependabot-maven.js';
 export { default as PomStorage } from './pom-store.js';
+export * from './pom-file-sort.js';
 export * from './pom-store.js';
